requirements.txt
python-telegram-bot==20.8
google-generativeai==0.3.2
httpx==0.25.2
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
colorlog==6.8.0